Add ability to setup automatic connection to a predefined SSID at boot
