+++ configs/config_default
@@ -185,6 +185,7 @@ CONFIG_ENV_DEFAULT_HOSTNAME="frankenstein"
 CONFIG_ENV_DEFAULT_STATION_IP="192.168.0.123"
 CONFIG_ENV_DEFAULT_STATION_MASK="255.255.255.0"
 CONFIG_ENV_DEFAULT_STATION_GW="192.168.0.1"
+# CONFIG_ENV_DEFAULT_STATION_AUTO_CONNECT is not set
 CONFIG_ENV_DEFAULT_AP_IP="192.168.1.1"
 CONFIG_ENV_DEFAULT_AP_MASK="255.255.255.0"
 CONFIG_ENV_DEFAULT_AP_GW="192.168.1.1"
