+++ include/iwconnect.h
@@ -1,0 +1,10 @@
+#ifndef IW_H__
+#define IW_H__
+
+extern int exec_iwconnect(const char *ssid, const char *password);
+
+#endif
+
+
+
+
